red test for sliding window day 1
