--- internal/pkg/problemsolvers/problemsolver.go
@@ -1,6 +1,0 @@
-package problemsolvers
-
-type ProblemSolver interface {
-	Initialize(initData string) error
-	Solve() (string, error)
-}
